Refactor: @Transactional 및 Service 분리
